+++ CHANGELOG.md
@@ -5,7 +5,8 @@
    * Added support for relative paths in `getLevelListing` action
 1. [](#bugfix)
    * Fixed admin not working with types that do not implement `FlexAuthorizeInterface`
-   * Bad redirect when changing parent of new page and saving [grav-plugin-admin#2014](https://github.com/getgrav/grav-plugin-admin/issues/2014)
+   * Fixed bad redirect when creating new flex object and choosing to create another return to the list
+   * Fixed bad redirect when changing parent of new page and saving [grav-plugin-admin#2014](https://github.com/getgrav/grav-plugin-admin/issues/2014)
    * Fixed page forms being empty if multi-language is enabled, but there's just one language [grav#3147](https://github.com/getgrav/grav/issues/3147)
    
 # v1.0.0-rc.20